@@ -14,19 +14,19 @@
   "private": true,
   "dependencies": {
     "@agm/core": "latest",
-    "@angular/animations": "^6.0.0",
-    "@angular/cdk": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
+    "@angular/animations": "^6.1.4",
+    "@angular/cdk": "^6.4.6",
+    "@angular/common": "^6.1.4",
+    "@angular/compiler": "^6.1.4",
+    "@angular/core": "^6.1.4",
     "@angular/flex-layout": "6.0.0-beta.15",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
-    "@angular/material": "^6.0.0",
-    "@angular/material-moment-adapter": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
+    "@angular/forms": "^6.1.4",
+    "@angular/http": "^6.1.4",
+    "@angular/material": "^6.4.6",
+    "@angular/material-moment-adapter": "^6.4.6",
+    "@angular/platform-browser": "^6.1.4",
+    "@angular/platform-browser-dynamic": "^6.1.4",
+    "@angular/router": "^6.1.4",
     "@google-cloud/functions-emulator": "^1.0.0-alpha.29",
     "@types/stripe-v3": "^3.0.7",
     "angularfire2": "^5.0.0-rc.10",
@@ -39,15 +39,15 @@
     "lodash": "latest",
     "moment": "latest",
     "ngx-bootstrap": "1.7.1",
-    "rxjs": "^6.1.0",
+    "rxjs": "^6.2.2",
     "rxjs-compat": "^6.1.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/core": "latest",
     "@angular/cli": "^6.0.7",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/language-service": "^6.0.0",
+    "@angular/compiler-cli": "^6.1.4",
+    "@angular/language-service": "^6.1.4",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/lodash": "latest",
@@ -64,7 +64,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.7.2",
+    "typescript": "2.9.2",
     "@angular-devkit/build-angular": "~0.6.0"
   }
 }
