@@ -1,12 +1,14 @@
-import {switchMap} from 'rxjs/operators';
-import {AfterViewInit, Component, OnDestroy, OnInit, ViewEncapsulation} from '@angular/core';
+import {filter, switchMap} from 'rxjs/operators';
+import {AfterViewInit, Component, HostListener, OnDestroy, OnInit, ViewEncapsulation} from '@angular/core';
 import {FootballCampService} from '../../services/football-camp/football-camp.service';
-import {ActivatedRoute, NavigationEnd, Params, Router} from '@angular/router';
+import {ActivatedRoute, Params, Router, Scroll} from '@angular/router';
 import * as _ from 'lodash';
 import {FootballCamp} from '../../models/football-camp';
 import {Session} from '../../models/session';
 import {SessionService} from '../../services/session/session.service';
 import {Meta, Title} from '@angular/platform-browser';
+import {ViewportScroller} from '@angular/common';
+import {Observable} from 'rxjs';
 
 @Component({
   selector: 'football-camp-details',
@@ -15,6 +17,11 @@ import {Meta, Title} from '@angular/platform-browser';
   encapsulation: ViewEncapsulation.None
 })
 export class FootballCampDetailsComponent implements OnInit, OnDestroy, AfterViewInit {
+
+  @HostListener("window:scroll", [])
+  onWindowScroll() {
+    console.log("Scroll Event");
+  }
 
   footballCamp: FootballCamp = null;
 
@@ -25,6 +32,7 @@ export class FootballCampDetailsComponent implements OnInit, OnDestroy, AfterVie
   fragment = '';
 
   constructor(private router: Router,
+              private viewportScroller: ViewportScroller,
               private route: ActivatedRoute,
               private footballCampService: FootballCampService,
               private sessionService: SessionService,
@@ -57,16 +65,37 @@ export class FootballCampDetailsComponent implements OnInit, OnDestroy, AfterVie
         this.sessions = sessions;
       });
 
-    this.router.events.subscribe(s => {
-      if (s instanceof NavigationEnd) {
-        const tree = this.router.parseUrl(this.router.url);
-        if (tree.fragment) {
-          const element = document.querySelector('#' + tree.fragment);
-          if (element) {
-            element.scrollIntoView(true);
-          }
+    // this.router.events.subscribe(s => {
+    //   if (s instanceof NavigationEnd) {
+    //     const tree = this.router.parseUrl(this.router.url);
+    //     if (tree.fragment) {
+    //       const element = document.querySelector('#' + tree.fragment);
+    //       if (element) {
+    //         element.scrollIntoView(true);
+    //       }
+    //     }
+    //   }
+    // });
+
+
+    this.router.events.pipe(
+      filter(e => e instanceof Scroll)
+    ).subscribe((e: Scroll) => {
+      if (e.anchor) {
+        const element = document.querySelector('#' + e.anchor);
+        if (element) {
+          element.scrollIntoView(true);
         }
       }
+    });
+
+    const eventOptions = {passive: true};
+    Observable.fromEvent(window, 'scroll', eventOptions).subscribe(e => {
+      console.log('scroll !');
+    });
+
+    Observable.fromEvent(window, 'click').subscribe(e => {
+      console.log('click !');
     });
   }
 
